+++ src/main/java/tk/themcbros/uselessmod/blocks/MetalBlock.java
@@ -1,0 +1,18 @@
+package tk.themcbros.uselessmod.blocks;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IWorldReader;
+
+public class MetalBlock extends Block {
+
+    public MetalBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public boolean isBeaconBase(BlockState state, IWorldReader world, BlockPos pos, BlockPos beacon) {
+        return true;
+    }
+}